Create tutorial7.cpp
some of this code came from basic tutorials 1-4 on ogre3d.org.  much of the rest of the code i wrote 2/7/2017 to 2/9/2017. when compiled a ninja stands on a hilly terrain with an ogre head revolving around him.  this was created for a course i took at kent state university.